Refactor model implementations (#225)
* refactor CFlow implementation

* refactor DFM implementation

* refactor PADIM implementation

* refactor PatchCore implementation

* refactor STFPM implementation

* revert model tests

* remove unintentionally committed file

* model.py -> lightning_model.py
