+++ anomalib/models/ganomaly/lightning_model.py
