+++ anomalib/models/patchcore/lightning_model.py
@@ -1,0 +1,108 @@
+"""Towards Total Recall in Industrial Anomaly Detection.
+
+Paper https://arxiv.org/abs/2106.08265.
+"""
+
+# Copyright (C) 2020 Intel Corporation
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions
+# and limitations under the License.
+
+from typing import List
+
+import torch
+from torch import Tensor
+
+from anomalib.models.components import AnomalyModule
+from anomalib.models.patchcore.torch_model import PatchcoreModel
+
+
+class PatchcoreLightning(AnomalyModule):
+    """PatchcoreLightning Module to train PatchCore algorithm.
+
+    Args:
+        layers (List[str]): Layers used for feature extraction
+        input_size (Tuple[int, int]): Input size for the model.
+        tile_size (Tuple[int, int]): Tile size
+        tile_stride (int): Stride for tiling
+        backbone (str, optional): Pre-trained model backbone. Defaults to "resnet18".
+        apply_tiling (bool, optional): Apply tiling. Defaults to False.
+    """
+
+    def __init__(self, hparams) -> None:
+        super().__init__(hparams)
+
+        self.model: PatchcoreModel = PatchcoreModel(
+            layers=hparams.model.layers,
+            input_size=hparams.model.input_size,
+            tile_size=hparams.dataset.tiling.tile_size,
+            tile_stride=hparams.dataset.tiling.stride,
+            backbone=hparams.model.backbone,
+            apply_tiling=hparams.dataset.tiling.apply,
+        )
+        self.embeddings: List[Tensor] = []
+
+    def configure_optimizers(self) -> None:
+        """Configure optimizers.
+
+        Returns:
+            None: Do not set optimizers by returning None.
+        """
+        return None
+
+    def training_step(self, batch, _batch_idx):  # pylint: disable=arguments-differ
+        """Generate feature embedding of the batch.
+
+        Args:
+            batch (Dict[str, Any]): Batch containing image filename, image, label and mask
+            _batch_idx (int): Batch Index
+
+        Returns:
+            Dict[str, np.ndarray]: Embedding Vector
+        """
+        self.model.feature_extractor.eval()
+        embedding = self.model(batch["image"])
+
+        # NOTE: `self.embedding` appends each batch embedding to
+        #   store the training set embedding. We manually append these
+        #   values mainly due to the new order of hooks introduced after PL v1.4.0
+        #   https://github.com/PyTorchLightning/pytorch-lightning/pull/7357
+        self.embeddings.append(embedding)
+
+    def on_validation_start(self) -> None:
+        """Apply subsampling to the embedding collected from the training set."""
+        # NOTE: Previous anomalib versions fit subsampling at the end of the epoch.
+        #   This is not possible anymore with PyTorch Lightning v1.4.0 since validation
+        #   is run within train epoch.
+        print("Aggregating the embedding extracted from the training set.")
+        embeddings = torch.vstack(self.embeddings)
+
+        sampling_ratio = self.hparams.model.coreset_sampling_ratio
+        self.model.subsample_embedding(embeddings, sampling_ratio)
+
+    def validation_step(self, batch, _):  # pylint: disable=arguments-differ
+        """Get batch of anomaly maps from input image batch.
+
+        Args:
+            batch (Dict[str, Any]): Batch containing image filename,
+                                    image, label and mask
+            _ (int): Batch Index
+
+        Returns:
+            Dict[str, Any]: Image filenames, test images, GT and predicted label/masks
+        """
+
+        anomaly_maps, anomaly_score = self.model(batch["image"])
+        batch["anomaly_maps"] = anomaly_maps
+        batch["pred_scores"] = anomaly_score.unsqueeze(0)
+
+        return batch
